feat: similar card improved
