+++ src/components/packagespage/package-card/home-page-package.tsx
@@ -197,7 +197,6 @@ export const SliderComponent = ({
   isOverlayVisible: boolean;
 }) => {
   const attr = pkg?.attributes;
-  console.log(attr);
   return (
     <Swiper
       spaceBetween={30}
